Change balance from u16 to i32

+++ src/network/packet/req_resp_balance.c
@@ -7,7 +7,7 @@ void pkt_make_req_balance(pkt_req_balance_t* pkt, u16 id) {
 	pkt->id = id;
 }
 
-void pkt_make_resp_balance(pkt_resp_balance_t* pkt, u16 balance) {
+void pkt_make_resp_balance(pkt_resp_balance_t* pkt, i32 balance) {
 	pkt_make_header(&pkt->header, RESP_BALANCE, PKT_RESP_BALANCE_PAYLOAD_SIZE);
 	pkt->balance = balance;
 }
